requirements.txt
Django
djangorestframework
mypy
django-stubs
djangorestframework-stubs